Split layout bundles into separate files
Move the layouts for each breakpoint into their own files. This makes it
possible for people to specify custom breakpoints when using `@import`
to include the component in the app's CSS. The pattern is as follows:

`@import "path/to/file.css" (min-width:20em);`

…where any valid Media Query can appear after the file path. This keeps
the breakpoints separate from the layout and is suitable for
optimization by build tools. It also makes it easier to concatenate the
CSS without any Media Queries and serve that up to IE 8.